added billy + joe + finalized syncing
