+++ app/controllers/admin/certification/ysws_controller.rb
@@ -166,32 +166,35 @@ class Admin::Certification::YswsController < Admin::Certification::ApplicationCo
   end
 
   def complete
-    @review = ::Certification::Ysws.find(params[:id])
+    Rails.logger.info "[YSWS#complete] user=#{current_user&.id} review=#{params[:id]} Starting complete action"
+
+    @review = ::Certification::Ysws.includes(:devlog_reviews).find(params[:id])
     authorize @review, :update?
 
-    # Run the Airtable sync synchronously so we know if it succeeds
-    # Only mark as reviewed if the sync succeeds
-    ActiveRecord::Base.transaction do
-      # Set reviewer_id to track who completed this review
-      @review.update!(reviewer_id: current_user.id)
-
-      # Perform sync synchronously - this will raise an error if it fails
-      ::Certification::Ysws::AirtableSyncJob.new.perform(@review.id)
-
-      # Only mark as reviewed if sync succeeded
-      @review.update!(reviewed_at: Time.current)
+    incomplete = @review.devlog_reviews.select { |dr| dr.pending? || dr.justification.blank? }
+    if incomplete.any?
+      Rails.logger.warn "[YSWS#complete] user=#{current_user&.id} review=#{params[:id]} Blocked: #{incomplete.count} incomplete devlog(s): #{incomplete.map(&:id).inspect}"
+      return render json: { success: false, error: "Fill in all devlogs" }, status: :unprocessable_entity
     end
+
+    @review.update_columns(reviewer_id: current_user.id, reviewed_at: Time.current)
+    Rails.logger.info "[YSWS#complete] user=#{current_user&.id} review=#{params[:id]} Marked reviewed_at=#{@review.reviewed_at}; enqueuing AirtableSyncJob"
+
+    ::Certification::YswsAirtableSyncJob.perform_later(@review.id)
+    Rails.logger.info "[YSWS#complete] user=#{current_user&.id} review=#{params[:id]} AirtableSyncJob enqueued successfully"
 
     render json: {
       success: true,
-      message: "Review completed and synced to Airtable successfully",
+      message: "Review completed! Syncing to Airtable in the background...",
       redirect_url: admin_certification_ysws_reviews_path
     }, status: :ok
   rescue StandardError => e
-    Sentry.capture_exception(e, tags: { category: 'certification.ysws' }, extra: { ysws_review_id: params[:id], user_id: current_user.id })
+    skip_authorization unless pundit_policy_authorized?
+    Rails.logger.error "[YSWS#complete] user=#{current_user&.id} review=#{params[:id]} #{e.class}: #{e.message}\n#{e.backtrace&.first(5)&.join("\n")}"
+    Sentry.capture_exception(e, tags: { category: 'certification.ysws' }, extra: { ysws_review_id: params[:id], user_id: current_user&.id })
     render json: {
       success: false,
-      error: "Failed to complete review and sync to Airtable: #{e.message}. Let AVD know!"
+      error: "Failed to complete review: #{e.message}. Let AVD know!"
     }, status: :unprocessable_entity
   end
 end
