+++ app/javascript/controllers/certification/ysws/complete_review_controller.js
@@ -7,7 +7,6 @@ export default class extends Controller {
   async complete(event) {
     event.preventDefault();
 
-    // Confirm with the reviewer
     if (
       !confirm(
         "Are you sure you want to complete this review? This will sync the review to Airtable and mark it as done.",
@@ -16,14 +15,11 @@ export default class extends Controller {
       return;
     }
 
-    // Disable the button while processing
     this.buttonTarget.disabled = true;
     this.buttonTarget.textContent = "Completing...";
 
     try {
-      const csrfToken = document.querySelector(
-        'meta[name="csrf-token"]',
-      ).content;
+      const csrfToken = document.querySelector('meta[name="csrf-token"]')?.content;
       const response = await fetch(
         `/admin/certification/review/${this.reviewIdValue}/complete`,
         {
@@ -38,30 +34,41 @@ export default class extends Controller {
       const data = await response.json();
 
       if (response.ok) {
-        // Show success message
-        alert(
-          data.message ||
-            "Review completed successfully! Redirecting to review queue...",
-        );
-
-        // Redirect to the review queue
-        if (data.redirect_url) {
-          window.location.href = data.redirect_url;
-        } else {
-          window.location.href = "/admin/certification/review";
-        }
+        this.showFlash(data.message || "Review completed successfully! Redirecting to review queue...", "success");
+        window.location.href = data.redirect_url || "/admin/certification/review";
       } else {
         const errorMessage =
           data.error || data.errors?.join(", ") || "Failed to complete review";
-        alert(`Error: ${errorMessage}`);
+        this.showFlash(errorMessage, "error");
         this.buttonTarget.disabled = false;
         this.buttonTarget.textContent = "Complete Review";
       }
     } catch (error) {
       console.error("Error completing review:", error);
-      alert("An unexpected error occurred. Please try again.");
+      this.showFlash("An unexpected error occurred. Please try again.", "error");
       this.buttonTarget.disabled = false;
       this.buttonTarget.textContent = "Complete Review";
     }
+  }
+
+  showFlash(message, variant = "error") {
+    let container = document.querySelector(".flash-container");
+    if (!container) {
+      container = document.createElement("div");
+      container.className = "flash-container";
+      document.body.appendChild(container);
+    }
+
+    const el = document.createElement("div");
+    el.className = `alert alert-${variant}`;
+    el.setAttribute("role", "alert");
+    el.setAttribute("aria-live", "assertive");
+    el.setAttribute("data-controller", "flash");
+    el.setAttribute("data-flash-timeout-value", "5000");
+    el.innerHTML = `
+      <div class="alert__content">${message}</div>
+      <button type="button" class="alert__close" aria-label="Close" data-action="click->flash#close">×</button>
+    `;
+    container.appendChild(el);
   }
 }
