+++ app/models/certification/ship.rb
@@ -9,23 +9,27 @@
 #  feedback         :text
 #  internal_reason  :text
 #  lock_version     :integer          default(0), not null
+#  recert_reason    :text
 #  stardust_earned  :integer
 #  status           :integer          default("pending"), not null
 #  created_at       :datetime         not null
 #  updated_at       :datetime         not null
 #  project_id       :bigint           not null
+#  returned_by_id   :bigint
 #  reviewer_id      :bigint
 #
 # Indexes
 #
-#  idx_on_status_claim_expires_at_c7a5e87a52        (status,claim_expires_at)
-#  index_certification_ship_reviews_on_decided_at   (decided_at)
-#  index_certification_ship_reviews_on_reviewer_id  (reviewer_id)
-#  index_ship_reviews_unique_pending_project        (project_id) UNIQUE WHERE (status = 0)
+#  idx_on_status_claim_expires_at_c7a5e87a52           (status,claim_expires_at)
+#  index_certification_ship_reviews_on_decided_at      (decided_at)
+#  index_certification_ship_reviews_on_returned_by_id  (returned_by_id)
+#  index_certification_ship_reviews_on_reviewer_id     (reviewer_id)
+#  index_ship_reviews_unique_pending_project           (project_id) UNIQUE WHERE (status = 0)
 #
 # Foreign Keys
 #
 #  fk_rails_...  (project_id => projects.id)
+#  fk_rails_...  (returned_by_id => users.id)
 #  fk_rails_...  (reviewer_id => users.id)
 #
 module Certification
