+++ app/views/admin/certification/ysws/show.html.erb
@@ -13,8 +13,14 @@
         <h1>YSWS Review</h1>
         <div style="display: flex; align-items: center; gap: var(--space-m);">
           <h2 class="project-name"><%= @review.project.title %></h2>
-          <% if @review.reviewed_at %>
-            <span class="status-done">done</span>
+          <% if @review.airtable_synced_at.present? %>
+            <% if @review.approved_minutes.to_i > 0 %>
+              <span class="status-approved">approved</span>
+            <% else %>
+              <span class="status-rejected">rejected</span>
+            <% end %>
+          <% elsif @review.reviewed_at.present? %>
+            <span class="status-awaiting-sync">awaiting sync</span>
           <% else %>
             <span class="status-pending">pending</span>
           <% end %>
@@ -177,8 +183,8 @@
             <span class="detail-value" style="display: flex; align-items: center; gap: var(--space-xs); flex-wrap: wrap;">
               <span><%= @review.user.display_name || @review.user.email || "Unknown User" %></span>
               <span class="fraud-links" style="display: inline-flex; gap: var(--space-xxs);">
-                <a href="#" class="fraud-link" target="_blank">Billy</a>
-                <a href="#" class="fraud-link" target="_blank">Joe</a>
+                <a href="https://billy.3kh0.net/?u=<%= @review.user.slack_id %>" class="fraud-link" target="_blank">Billy</a>
+                <a href="https://joe.fraud.hackclub.com/profile/<%= @review.user.slack_id %>" class="fraud-link" target="_blank">Joe</a>
               </span>
             </span>
           </div>
